+++ extensions/word-search/src/hooks/use-suggestions.ts
@@ -4,31 +4,29 @@ import { URL, URLSearchParams } from "url";
 import { Toast, showToast } from "@raycast/api";
 import { useFetch } from "@raycast/utils";
 
-import type { SearchType, Word } from "@/types";
+import type { SimpleWord } from "@/types";
 import { Vocabulary } from "@/types";
 
 import { useCapitalizeResults, useMaxResults } from "@/hooks/use-settings";
 
-const useSearchWords = (wordsToSearch: string, type: SearchType, vocabulary?: Vocabulary) => {
+const useSuggestions = (wordsToSearch: string, vocabulary?: Vocabulary) => {
   const capitalizeResults = useCapitalizeResults();
   const maxResults = useMaxResults();
 
   const url = useMemo(() => {
     const searchParams = new URLSearchParams({
-      md: "d",
       max: maxResults.toString(),
-      [type]: wordsToSearch,
+      s: wordsToSearch,
     });
 
-    // TODO: Make sure to remove this for the wrong types (See https://www.datamuse.com/api/#vocabs, this only worls for ml, sl, sp)
     if (vocabulary && vocabulary !== Vocabulary.English) {
       searchParams.set("v", vocabulary);
     }
 
-    return new URL(`/words?${searchParams}`, "https://api.datamuse.com/words").toString();
-  }, [wordsToSearch, type, maxResults, vocabulary]);
+    return new URL(`/sug?${searchParams}`, "https://api.datamuse.com/sug").toString();
+  }, [wordsToSearch, maxResults, vocabulary]);
 
-  return useFetch<Word[]>(url, {
+  return useFetch<SimpleWord[]>(url, {
     parseResponse: async (response) => {
       if (!response.ok) {
         await showToast(
@@ -39,21 +37,11 @@ const useSearchWords = (wordsToSearch: string, type: SearchType, vocabulary?: Vo
         return [];
       }
 
-      const words = (await response.json()) as Word[];
+      const words = (await response.json()) as SimpleWord[];
 
       words.forEach((word) => {
-        if (word.defs == undefined || !word.defs.length) {
-          return;
-        }
-
         if (capitalizeResults && wordsToSearch[0] === wordsToSearch[0].toUpperCase()) {
           word.word = word.word[0].toUpperCase() + word.word.slice(1);
-        }
-
-        for (let i = 0; i < word.defs.length; i++) {
-          let definition: string = word.defs[i];
-          definition = word.defs[i].replace(/\t/g, "~");
-          word.defs[i] = definition.split("~")[1];
         }
       });
 
@@ -63,4 +51,4 @@ const useSearchWords = (wordsToSearch: string, type: SearchType, vocabulary?: Vo
   });
 };
 
-export default useSearchWords;
+export default useSuggestions;