+++ extensions/word-search/src/word-adjective-search.tsx
@@ -1,7 +1,8 @@
+import type { LaunchProps } from "@raycast/api";
+
 import { SearchType } from "@/types";
 
 import SearchResults from "@/components/SearchResults";
-import { LaunchProps } from "@raycast/api";
 
 export default function SearchAdjective(props: LaunchProps) {
   return SearchResults(SearchType.ADJECTIVE, "Search for adjectives that describe a word", props);