+++ extensions/word-search/src/word-antonym-search.tsx
@@ -1,7 +1,8 @@
+import type { LaunchProps } from "@raycast/api";
+
 import { SearchType } from "@/types";
 
 import SearchResults from "@/components/SearchResults";
-import {LaunchProps} from '@raycast/api';
 
 export default function SearchAntonyms(props: LaunchProps) {
   return SearchResults(SearchType.ANTONYM, "Search for antonyms", props);