+++ extensions/word-search/src/word-sound-like-search.tsx
@@ -1,7 +1,8 @@
+import type { LaunchProps } from "@raycast/api";
+
 import { SearchType } from "@/types";
 
 import SearchResults from "@/components/SearchResults";
-import {LaunchProps} from '@raycast/api';
 
 export default function SearchSynonym(props: LaunchProps) {
   return SearchResults(SearchType.SOUND_LIKE, "Search for words that sound like", props);