+++ extensions/word-search/src/word-spell.tsx
@@ -1,7 +1,8 @@
+import type { LaunchProps } from "@raycast/api";
+
 import { SearchType } from "@/types";
 
 import SearchResults from "@/components/SearchResults";
-import {LaunchProps} from '@raycast/api';
 
 export default function SearchSpelling(props: LaunchProps) {
   return SearchResults(SearchType.MISSING_LETTERS, "Spell a word", props);